refactor(UI): relocated services
