+++ ScriptBeeClient/src/app/services/common/client-id.service.ts
