+++ ScriptBeeClient/src/app/services/common/download.service.ts
