+++ ScriptBeeClient/src/app/utils/client-id.interceptor.ts
@@ -1,6 +1,6 @@
 import { HttpInterceptorFn } from '@angular/common/http';
 import { inject } from '@angular/core';
-import { ClientIdService } from '../services/client-id.service';
+import { ClientIdService } from '../services/common/client-id.service';
 
 export const clientIdInterceptor: HttpInterceptorFn = (req, next) => {
   const clientIdService = inject(ClientIdService);
